Allow sips URI with transport=tcp
Scenario: Oracle SBC sends us an INVITE with a Contact that
specifies both sips and "transport=tcp". If we try to
issue a re-INVITE on this session, it fails, generating a
"416 Invalid URI" error, presumambly because of the conflict
between "sips" and "transport=tls"

From RFC3261, Section 26.2.2 <https://tools.ietf.org/html/rfc3261#section-26.2.2>

     Note that in the SIPS URI scheme, transport is independent of TLS,
     and thus "sips:alice@atlanta.com;transport=TCP" and
     "sips:alice@atlanta.com;transport=sctp" are both valid (although
     note that UDP is not a valid transport for SIPS).  The use of
     "transport=tls" has consequently been deprecated, partly because
     it was specific to a single hop of the request.  This is a change
     since RFC 2543 <https://tools.ietf.org/html/rfc2543>.

In nta_tpn_by_url, add a specific test for scheme="sips"
and "transport=tls": if this occurs, override and set the
protocol to "tls".